requirements.txt
absl-py==0.7.1
astor==0.8.0
botocore==1.12.86
cement==2.8.2
certifi==2019.6.16
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
cython==0.29.13
docutils==0.14
Flask==1.0.3
Flask-Cors==3.0.8
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
labelImg==1.8.3
lxml==4.4.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
mock==3.0.5
numpy==1.16.4
opencv-python==4.1.0.25
pathspec==0.5.5
pillow==6.1.0
progressbar==2.5
protobuf==3.9.0
pycocotools==2.0
pyparsing==2.4.2
PyQt5==5.13.0
PyQt5-sip==4.19.18
pyreadline==2.1
python-dateutil==2.8.0
PyYAML==3.13
requests==2.22.0
scipy==1.3.0
semantic-version==2.5.0
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow==1.14.0
termcolor==1.1.0
urllib3==1.25.3
Werkzeug==0.15.4
wincertstore==0.2
wrapt==1.11.2
